pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iori-spec"
version = "0.1.0"
description = "Static analysis toolkit for specs: SSoT, traceability, and vocabulary consistency."
authors = [{ name = "YOUR_NAME", email = "you@example.com" }]
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
iori-spec = "iori_spec.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
